modify
git-svn-id: https://svn.baidu.com/app/search/sep/trunk/anti-spam/gezi@7019 fa64baa9-71d1-4fed-97ae-c15534abce97

+++ include/debug_util.h
@@ -130,21 +130,21 @@ void PRange(Iter begin, Iter end, std::ostream& out = std::cout, const string& s
 
 #define PVEC(v)\
         VLOG(5) << #v <<" --- " << v.size();\
-        for (int i = 0; i < v.size(); i++)\
+        for (size_t i = 0; i < v.size(); i++)\
         {\
             VLOG(5) << setiosflags(ios::left) << setfill(' ') << setw(10) << i << v[i] << "\n";\
         }
 
 #define PVEC_TOPN(v, n)\
 	VLOG(5) << #v <<" --- " << v.size();\
-	for (int i = 0; i < v.size() && i < n; i++)\
+	for (size_t i = 0; i < v.size() && i < n; i++)\
 {\
 	VLOG(5) << setiosflags(ios::left) << setfill(' ') << setw(10) << i << v[i] << "\n";\
 }
 
 #define Pvec(v)\
         DLOG(INFO) << #v <<" --- " << v.size();\
-        for (int i = 0; i < v.size(); i++)\
+        for (size_t i = 0; i < v.size(); i++)\
         {\
             DLOG(INFO) << setiosflags(ios::left) << setfill(' ') << setw(10) << i << v[i] << "\n";\
         }
